refactor(TranslateVariants): split core.ts and constant.ts from main file
