+++ src/TranslateVariants/LICENSE
@@ -1,8 +1,7 @@
 /**
  * SPDX-License-Identifier: CC-BY-SA-4.0
- * _addText: '{{Gadget Header|license=CC-BY-SA-4.0}}'
+ * _addText: '{{Gadget Header|url=https://git.qiuwen.net.cn/InterfaceAdmin/QiuwenGadgets/src/branch/master/src/TranslateVariants|license=CC-BY-SA-4.0}}'
  *
  * @base <https://zh.wikipedia.org/wiki/User:Xiplus/js/TranslateVariants>
  * @source <https://git.qiuwen.net.cn/InterfaceAdmin/QiuwenGadgets/src/branch/master/src/TranslateVariants>
- * @dependency ext.gadget.i18n, mediawiki.api, mediawiki.diff.styles, mediawiki.util
  */
