+++ src/TranslateVariants/modules/constant.ts
@@ -1,0 +1,2 @@
+export const TRANSLATE_VARIANTS_SUMMARY =
+	'自动转换变体自[[$1]]（[[MediaWiki:Gadget-TranslateVariants.js|TranslateVariants]]）';
